+++ docker/startup.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+exec bundle exec rails s -b 0.0.0.0