+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Set env `FIREBASE_DYNAMIC_LINK_DOMAIN` to Firebase's dynamic link domain
   - Set env `FIREBASE_ANDROID_PACKAGE` to the android package of mudamos mobile
   - Set env `FIREBASE_IOS_PACKAGE` to the ios package of mudamos mobile
+  - Set env `PETITION_SHARE_LINK_GENERATION_QUEUE` to the create share link queue
   - Run `rake db:migrate`
   - Run `rake petitions:generate_share_links`
 * [PR 447]: Add API call to get one plip by slug